agregado sin correos y color no leido
