+++ php/contacto/contacto_correo.php
@@ -1,0 +1,11 @@
+<?php
+
+
+
+  function correo_abrir($id,$nombre,$correo,$motivo,$comentario,$fecha){
+    print ('
+
+
+    ');
+  }
+?>
